+++ smart-contracts/contracts-common/concordium-contracts-common/src/types.rs
@@ -516,8 +516,12 @@ impl FromStr for PublicKeyEcdsaSecp256k1 {
 
 /// Signature for a Ed25519 message. Must be 64 bytes long.
 #[derive(Copy, Clone, Debug, PartialEq, PartialOrd, Eq, Ord, crate::Deserial, crate::Serial)]
+#[cfg_attr(feature = "derive-serde", derive(serde::Deserialize, serde::Serialize))]
 #[repr(transparent)]
-pub struct SignatureEd25519(pub [u8; 64]);
+pub struct SignatureEd25519(
+    #[cfg_attr(feature = "derive-serde", serde(with = "serde_impl::fixed_byte_array_64_hex"))]
+    pub  [u8; 64],
+);
 
 impl fmt::Display for SignatureEd25519 {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
@@ -619,6 +623,7 @@ pub struct AccountPublicKeys {
 pub(crate) type CredentialIndex = u8;
 
 #[derive(crate::Serialize, Debug, SchemaType)]
+#[cfg_attr(feature = "derive-serde", derive(serde::Deserialize, serde::Serialize))]
 #[non_exhaustive]
 /// A cryptographic signature indexed by the signature scheme. Currently only a
 /// single scheme is supported, `ed25519`.
@@ -627,6 +632,7 @@ pub enum Signature {
 }
 
 #[derive(crate::Serialize, Debug, SchemaType)]
+#[cfg_attr(feature = "derive-serde", derive(serde::Deserialize, serde::Serialize))]
 #[concordium(transparent)]
 /// Account signatures. This is an analogue of transaction signatures that are
 /// part of transactions that get sent to the chain.
@@ -639,6 +645,7 @@ pub struct AccountSignatures {
 }
 
 #[derive(crate::Serialize, Debug, SchemaType)]
+#[cfg_attr(feature = "derive-serde", derive(serde::Deserialize, serde::Serialize))]
 #[concordium(transparent)]
 pub struct CredentialSignatures {
     #[concordium(size_length = 1)]
@@ -2693,6 +2700,46 @@ mod serde_impl {
     impl<'de> SerdeDeserialize<'de> for AccountAddress {
         fn deserialize<D: Deserializer<'de>>(des: D) -> Result<Self, D::Error> {
             des.deserialize_str(Base58Visitor)
+        }
+    }
+
+    /// Helper for [de]serializing a fixed length array of 64 as an hex string.
+    pub(super) mod fixed_byte_array_64_hex {
+        use super::*;
+
+        /// Serialize (via Serde)
+        pub fn serialize<S: serde::Serializer>(dt: &[u8], ser: S) -> Result<S::Ok, S::Error> {
+            ser.serialize_str(hex::encode(dt).as_str())
+        }
+
+        /// Deserialize (via Serde)
+        pub fn deserialize<'de, D: serde::Deserializer<'de>>(des: D) -> Result<[u8; 64], D::Error> {
+            struct HexVisitor;
+            impl<'de> serde::de::Visitor<'de> for HexVisitor {
+                type Value = [u8; 64];
+
+                fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+                    write!(formatter, "A hex string.")
+                }
+
+                fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+                where
+                    E: serde::de::Error, {
+                    let r = hex::decode(v).map_err(serde::de::Error::custom)?;
+
+                    if r.len() != 64 {
+                        return Err(serde::de::Error::invalid_length(
+                            r.len(),
+                            &"Length of 64 was expected.",
+                        ));
+                    }
+
+                    let mut array = [0u8; 64];
+                    array.copy_from_slice(&r);
+                    Ok(array)
+                }
+            }
+            des.deserialize_str(HexVisitor)
         }
     }